Implement shared state struct and implement review suggestions
