--- server/src/slicec_ext/mod.rs
@@ -1,3 +1,0 @@
-// Copyright (c) ZeroC, Inc.
-
-pub mod diagnostic_ext;
